Feature: Basic navigation now satisifying. Next is toolbar, and fixing toast messages.
