+++ app/build.gradle
@@ -3,6 +3,8 @@ apply plugin: 'kotlin-android'
 
 apply plugin: 'kotlin-kapt'
 
+apply plugin: 'androidx.navigation.safeargs.kotlin'
+
 android {
     compileSdkVersion 30
 
@@ -48,7 +50,7 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 
     // Fragment
-    implementation "androidx.fragment:fragment-ktx:1.3.0"
+    implementation "androidx.fragment:fragment-ktx:1.3.1"
 
     // Recyclerview
     implementation "androidx.recyclerview:recyclerview:1.1.0"
@@ -61,17 +63,19 @@ dependencies {
     androidTestImplementation "androidx.room:room-testing:$room_version"
 
     // Navigation
-    def nav_version = '2.3.3'
+    def nav_version = '2.3.4'
     implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
     implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
     implementation "androidx.navigation:navigation-dynamic-features-fragment:$nav_version"
 
     // Lifecycle
-
+    implementation 'androidx.lifecycle:lifecycle-process:2.3.0'
+    implementation 'androidx.lifecycle:lifecycle-service:2.3.0'
     implementation "androidx.lifecycle:lifecycle-viewmodel-savedstate:2.3.0"
 
 
     // Coroutines
     api "org.jetbrains.kotlinx:kotlinx-coroutines-core:$rootProject.coroutines"
     api "org.jetbrains.kotlinx:kotlinx-coroutines-android:$rootProject.coroutines"
+
 }
