+++ app/src/main/AndroidManifest.xml
@@ -3,20 +3,28 @@
     package="sebgg.dev.workoutdiary">
 
     <application
-        android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.WorkoutDiary"
+        android:theme="@style/Theme.MaterialComponents.DayNight.NoActionBar"
         android:name=".WorkoutApplication">
-        <activity android:name=".MainActivity">
+        <activity android:name=".activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+<!--        <activity android:name=".activities.HistoryActivity"-->
+<!--            android:label="Workout History"-->
+<!--            android:parentActivityName=".activities.MainActivity">-->
+
+<!--            <meta-data-->
+<!--                android:name="android.support.PARENT_ACTIVITY"-->
+<!--                android:value=".activities.MainActivity"/>-->
+<!--        </activity>-->
     </application>
 
 </manifest>
